Handle JAVA_OPTS the same way in Java Main as all other containers
Previously, Java Main handled JAVA_OPTS differently than all other
containers.  Specifically, it added the opts directly to the command
line rather than first placing them in an environment variable and
then referencing that variable.  This caused issues because you could
not escape a variable the same way for all containers.

This change causes a JAVA_OPTS environment variable to be written and
then referenced later in the command.

[#108647188][resolves #248]