Remove the user of expr
Previously, there were a number of places where we used $(expr ...) to
parse the contents the VCAP_APPLICATION environment variable at
runtime.  Unbeknownst to us expr has a special case where, if the
returned value of the expression is 0, it exits.  As we were using
expr to parse the instance_index, applications would start all of
their instances, except the 0th instance.  As you can imagine, this
took quite some time to detect.

This change dumps the use of $(expr ..) in favor of $(ruby -e ...).
While this now ties us to ruby being available in the environment, it
does give us some serious advantages.  First off, we can parse out the
0th instance_index.  Secondly we can now write much more
understandable expressions like
"#{a['application_name']}:#{a['instance_index']}".

[1]: http://unix.stackexchange.com/questions/63166/bash-e-exits-when-let-or-expr-evaluates-to-0