Container Customizer
With a recent update to the Spring Boot support in the buildpack, Spring Boot
WARs were broken.  The root cause is that the embedded Tomcat container would
not allow additional libraries to be symbolically linked to the WEB-INF
directory.  This change adds a framework that notices that the application is
a Spring Boot WAR and contributes a Tomcat LifecycleListener that adjusts this
configuration.